Manual session going through the states, marking code.
I probably marked a few switches as code, I'll fix those later if there are any. I also atuolabeled all the code from 00:0000 to 1D:1FFF, basically the entire ROM. It seems to have went well. This code dump is probably huge. But if you want to do your own diggint into this, the objects and states are a good place to do it.