Cross-bank gone through. Details inside.
It seems to actually be a map/game world mod routines. It seems to schedule tasks into slots that I haven't gotten to yet. And it is mostly used to queue the background updates and features related. Probably world palette updates, too. All are roughly commented, but not too complicated to go through.